+++ package.json
@@ -22,6 +22,7 @@
     "@tanstack/react-query": "^5.53.3",
     "@tanstack/react-query-devtools": "^5.53.3",
     "@tanstack/react-table": "^8.17.3",
+    "awesome-debounce-promise": "^2.1.0",
     "axios": "^1.7.2",
     "clsx": "^2.1.1",
     "dayjs": "^1.11.11",
