+++ src/lib/components/form/FormAsyncSelect.tsx
@@ -1,34 +1,17 @@
-import { type PostalCode, searchPostalCodes } from '@/lib/api';
-import { getCountryFromCode } from '@/pages-components/Dashboard/NewEmployeePage/const';
 import { type UseControllerProps, useController } from 'react-hook-form';
 
 import { AsyncSelectWithLabels, type AsyncSelectWithLabelsProps } from '@/ui/hocs';
+import type { AsyncProps } from 'react-select/async';
 
-function mapPostalCodes(postalCodes: PostalCode[]) {
-  return postalCodes.map((p) => ({
-    value: p.id,
-    label: `${p.postalCode}, ${p.city || p.region}, ${getCountryFromCode(p.countryCode).name}`,
-  }));
-}
-
-const promiseOptions = async (inputValue: string) => {
-  if (inputValue?.length < 3) {
-    return [];
-  }
-
-  const res = await searchPostalCodes(inputValue);
-  return mapPostalCodes(res);
-};
-
-interface FormAsyncSelectProps extends Omit<AsyncSelectWithLabelsProps, 'value' | 'onChange' | 'loadOptions'> {
+export interface FormAsyncSelectProps extends Omit<AsyncSelectWithLabelsProps, 'value' | 'onChange' | 'loadOptions'> {
   name: string;
   initialValue?: string;
   rules?: UseControllerProps['rules'];
-  // promiseOptions: AsyncProps<any, false, any>['loadOptions'];
+  promisedOptions: AsyncProps<any, false, any>['loadOptions'];
 }
 
 export const FormAsyncSelect: React.FC<FormAsyncSelectProps> = (props) => {
-  const { name, initialValue, rules, isDisabled, ...rest } = props;
+  const { name, initialValue, rules, isDisabled, promisedOptions, ...rest } = props;
   const {
     field: { value, onChange, onBlur },
     fieldState: { isTouched, isDirty, error },
@@ -40,7 +23,7 @@ export const FormAsyncSelect: React.FC<FormAsyncSelectProps> = (props) => {
       name={name}
       {...rest}
       value={value}
-      loadOptions={promiseOptions}
+      loadOptions={promisedOptions}
       errorText={(initialValue && error?.message) || ((isTouched && isDirty && error?.message) as string)}
       isDisabled={isSubmitting || isDisabled}
       onBlur={onBlur}
